+++ crates/core/src/models/balance.rs
@@ -1,0 +1,15 @@
+use reth_primitives::{Address, U256};
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct TokenBalance {
+    pub contract_address: Address,
+    pub token_balance: Option<U256>,
+    pub error: Option<String>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct TokenBalances {
+    pub address: Address,
+    pub token_balances: Vec<TokenBalance>,
+}
